+++ spec/unit/hanami/rspec/commands/install_spec.rb
@@ -25,10 +25,17 @@ RSpec.describe Hanami::RSpec::Commands::Install do
       # Gemfile
       gemfile = <<~EOF
         group :test do
+          gem "capybara"
           gem "rack-test"
         end
       EOF
       expect(fs.read("Gemfile")).to include(gemfile)
+
+      # .gitignore
+      gitignore = <<~EOF
+        spec/examples.txt
+      EOF
+      expect(fs.read(".gitignore")).to include(gitignore)
 
       # .rspec
       dotrspec = <<~EOF
@@ -47,6 +54,7 @@ RSpec.describe Hanami::RSpec::Commands::Install do
         require "hanami/prepare"
 
         require_relative "support/rspec"
+        require_relative "support/features"
         require_relative "support/requests"
       EOF
       expect(fs.read("spec/spec_helper.rb")).to eq(spec_helper)
@@ -56,32 +64,76 @@ RSpec.describe Hanami::RSpec::Commands::Install do
         # frozen_string_literal: true
 
         RSpec.configure do |config|
+          # Use the recommended non-monkey patched syntax.
+          config.disable_monkey_patching!
+
+          # Use and configure rspec-expectations.
           config.expect_with :rspec do |expectations|
+            # This option will default to `true` in RSpec 4.
             expectations.include_chain_clauses_in_custom_matcher_descriptions = true
           end
 
+          # Use and configure rspec-mocks.
           config.mock_with :rspec do |mocks|
+            # Prevents you from mocking or stubbing a method that does not exist on a
+            # real object.
             mocks.verify_partial_doubles = true
           end
 
+          # This option will default to `:apply_to_host_groups` in RSpec 4.
           config.shared_context_metadata_behavior = :apply_to_host_groups
 
+          # Limit a spec run to individual examples or groups you care about by tagging
+          # them with `:focus` metadata. When nothing is tagged with `:focus`, all
+          # examples get run.
+          #
+          # RSpec also provides aliases for `it`, `describe`, and `context` that include
+          # `:focus` metadata: `fit`, `fdescribe` and `fcontext`, respectively.
           config.filter_run_when_matching :focus
 
-          config.disable_monkey_patching!
-          config.warnings = true
+          # Allow RSpec to persist some state between runs in order to support the
+          # `--only-failures` and `--next-failure` CLI options. We recommend you
+          # configure your source control system to ignore this file.
+          config.example_status_persistence_file_path = "spec/examples.txt"
 
+          # Uncomment this to enable warnings. This is recommended, but in some cases
+          # may be too noisy due to issues in dependencies.
+          # config.warnings = true
+
+          # Show more verbose output when running an individual spec file.
           if config.files_to_run.one?
             config.default_formatter = "doc"
           end
 
+          # Print the 10 slowest examples and example groups at the end of the spec run,
+          # to help surface which specs are running particularly slow.
           config.profile_examples = 10
 
+          # Run specs in random order to surface order dependencies. If you find an
+          # order dependency and want to debug it, you can fix the order by providing
+          # the seed, which is printed after each run:
+          #
+          # --seed 1234
           config.order = :random
+
+          # Seed global randomization in this process using the `--seed` CLI option.
+          # This allows you to use `--seed` to deterministically reproduce test failures
+          # related to randomization by passing the same `--seed` value as the one that
+          # triggered the failure.
           Kernel.srand config.seed
         end
       EOF
       expect(fs.read("spec/support/rspec.rb")).to eq(support_rspec)
+
+      # spec/support/features.rb
+      support_features = <<~EOF
+        # frozen_string_literal: true
+
+        require "capybara/rspec"
+
+        Capybara.app = Hanami.app
+      EOF
+      expect(fs.read("spec/support/features.rb")).to eq(support_features)
 
       # spec/support/requests.rb
       support_requests = <<~EOF
@@ -89,13 +141,14 @@ RSpec.describe Hanami::RSpec::Commands::Install do
 
         require "rack/test"
 
-        RSpec.shared_context "Hanami app" do
+        RSpec.shared_context "Rack::Test" do
+          # Define the app for Rack::Test requests
           let(:app) { Hanami.app }
         end
 
         RSpec.configure do |config|
           config.include Rack::Test::Methods, type: :request
-          config.include_context "Hanami app", type: :request
+          config.include_context "Rack::Test", type: :request
         end
       EOF
       expect(fs.read("spec/support/requests.rb")).to eq(support_requests)