+++ spec/unit/hanami/rspec/version_spec.rb
@@ -2,6 +2,6 @@
 
 RSpec.describe "Hanami::RSpec::VERSION" do
   it "returns version" do
-    expect(Hanami::RSpec::VERSION).to eq("2.1.0.rc1")
+    expect(Hanami::RSpec::VERSION).to eq("2.1.0.rc2")
   end
 end