+++ CHANGELOG.md
@@ -6,6 +6,10 @@ RSpec support for Hanami
 
 - [Tim Riley] Install Capybara and generate `spec/support/capybara.rb` in `hanami install` hook.
 
+### Changed
+
+- [Tim Riley] Add explanatory code comments to `spec/support/rspec.rb` generated in `hanami install` hook.
+
 ## v2.1.0.rc1 - 2023-11-01
 
 ### Added