+++ CHANGELOG.md
@@ -4,6 +4,7 @@ RSpec support for Hanami
 
 ### Added
 
+- [Tim Riley] Skip generating tests for actions and parts when `--skip-tests` CLI flag is given.
 - [Tim Riley] Install Capybara and generate `spec/support/capybara.rb` in `hanami install` hook.
 
 ### Changed